Also make TDigest.count a uint type

+++ tdigest.go
@@ -55,7 +55,7 @@ func centroidLessOrEquals(p, q interface{}) bool {
 type TDigest struct {
 	summary     *avltree.Tree
 	compression uint
-	count       float64
+	count       uint
 }
 
 func New(compression uint) *TDigest {
@@ -75,7 +75,7 @@ func (t *TDigest) Percentile(p float64) float64 {
 		return t.summary.At(0).(Centroid).mean
 	}
 
-	p *= t.count
+	p *= float64(t.count)
 	var total float64 = 0
 	i := 0
 
@@ -99,7 +99,7 @@ func (t *TDigest) Percentile(p float64) float64 {
 }
 
 func (t *TDigest) Update(value float64, weight uint) {
-	t.count += float64(weight)
+	t.count += weight
 
 	newCentroid := Centroid{value, weight}
 
@@ -176,7 +176,7 @@ func (t *TDigest) Merge(other *TDigest) {
 }
 
 func (t TDigest) String() string {
-	return fmt.Sprintf("TD<compression=%d, count=%.1f, centroids=%d>", t.compression, t.count, t.summary.Len())
+	return fmt.Sprintf("TD<compression=%d, count=%d, centroids=%d>", t.compression, t.count, t.summary.Len())
 }
 
 func (t *TDigest) updateCentroid(c Centroid, mean float64, weight uint) {
@@ -190,7 +190,7 @@ func (t *TDigest) updateCentroid(c Centroid, mean float64, weight uint) {
 }
 
 func (t *TDigest) threshold(q float64) float64 {
-	return (4 * t.count * q * (1 - q)) / float64(t.compression)
+	return (4 * float64(t.count) * q * (1 - q)) / float64(t.compression)
 }
 
 func (t *TDigest) computeCentroidQuantile(c Centroid) float64 {
@@ -200,7 +200,7 @@ func (t *TDigest) computeCentroidQuantile(c Centroid) float64 {
 		cumSum += item.count
 	}
 
-	return (float64(c.count)/2.0 + float64(cumSum)) / t.count
+	return (float64(c.count)/2.0 + float64(cumSum)) / float64(t.count)
 }
 
 func (t *TDigest) addCentroid(c Centroid) {
